Move export logs to quick actions (to sit with report issue button)
